+++ Examples/Digit/DigitSingleStepRobustnessTest.cpp
@@ -22,7 +22,7 @@ int main(int argc, char* argv[]) {
     
     pinocchio::Model model_double;
     pinocchio::urdf::buildModel(urdf_filename, model_double);
-    pinocchio::ModelTpl<float> model = model_double.cast<float>();
+    pinocchio::ModelTpl<double> model = model_double.cast<double>();
 
     model.gravity.linear()(2) = GRAVITY;
     
@@ -44,7 +44,7 @@ int main(int argc, char* argv[]) {
     model.armature(model.getJointId("right_toe_B") - 1) = 0.036089475;
 
     // load settings
-    const float T = 0.4;
+    const double T = 0.4;
     const TimeDiscretization time_discretization = Chebyshev;
     int N = 16;
     int degree = 5;
@@ -85,7 +85,7 @@ int main(int argc, char* argv[]) {
 
                 for (int test_id = 1; test_id <= 100; test_id++) {
                     std::srand(test_id);
-                    Eigen::VectorXf z = 0.2 * Eigen::VectorXf::Random((degree + 1) * NUM_INDEPENDENT_JOINTS + NUM_JOINTS + NUM_DEPENDENT_JOINTS).array() - 0.1;
+                    Eigen::VectorXd z = 0.2 * Eigen::VectorXd::Random((degree + 1) * NUM_INDEPENDENT_JOINTS + NUM_JOINTS + NUM_DEPENDENT_JOINTS).array() - 0.1;
                     
                     SmartPtr<DigitSingleStepOptimizer> mynlp = new DigitSingleStepOptimizer();
                     try {
@@ -117,7 +117,7 @@ int main(int argc, char* argv[]) {
                         status = app->OptimizeTNLP(mynlp);
 
                         auto end = std::chrono::high_resolution_clock::now();
-                        float solve_time = std::chrono::duration_cast<std::chrono::milliseconds>(end - start).count() / 1000.0;
+                        double solve_time = std::chrono::duration_cast<std::chrono::milliseconds>(end - start).count() / 1000.0;
 
                         experiment_output << mynlp->obj_value_copy << ' ' << mynlp->final_constr_violation << ' ' << solve_time << std::endl;
                     }
